+++ src/components/index.js
@@ -1,5 +1,6 @@
 import Header from "./Header";
 import Keyboard from "./Keyboard";
 import Modal from "./Modal";
+import Grid from "./Grid";
 
-export { Header, Keyboard, Modal };
+export { Header, Keyboard, Modal, Grid };
